Correct HTML homework and add the last homework to C#1

+++ C#1/Loops/OddAndEvenProduct/OddAndEvenProduct.cs
@@ -5,8 +5,8 @@ class OddAndEvenProduct
     static void Main()
     {
         int n = int.Parse(Console.ReadLine());
-        int oddProduct = 1;
-        int evenProduct = 1;
+        long oddProduct = 1;
+        long evenProduct = 1;
         var numbers = Console.ReadLine().Split(' ').Select(int.Parse).ToArray();
         
         for (int i = 0; i <= n-1; i++)
